Split recall feedback tests for line lint
