+++ internal/http/control_portal_types.go
@@ -1,0 +1,36 @@
+package http
+
+import (
+	nethttp "net/http"
+
+	"github.com/markhuangai/dense-mem/internal/http/handler"
+	"github.com/markhuangai/dense-mem/internal/observability"
+	"github.com/markhuangai/dense-mem/internal/service"
+	"github.com/markhuangai/dense-mem/internal/service/dreamservice"
+)
+
+type ControlPortalTelemetry struct {
+	Reader         service.TelemetryReader
+	HTTPMetrics    observability.HTTPMetrics
+	ScrapeHandler  nethttp.Handler
+	ScrapeToken    string
+	SSO            *service.SSOService
+	Config         service.AppConfigService
+	Logs           service.OperationLogReader
+	RecallFeedback service.RecallFeedbackEventReader
+	Dreams         dreamservice.Service
+}
+
+type controlPortalHandler struct {
+	profiles       handler.ProfileServiceInterface
+	keys           handler.APIKeyServiceInterface
+	security       service.SecurityService
+	metrics        service.UsageMetricsReader
+	telemetry      service.TelemetryReader
+	operationLogs  service.OperationLogReader
+	recallFeedback service.RecallFeedbackEventReader
+	dreams         dreamservice.Service
+	health         HealthConfig
+	sso            *service.SSOService
+	appConfig      service.AppConfigService
+}
