+++ web/src/App.test.tsx
@@ -2,7 +2,7 @@ import { render, screen, waitFor, within } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { beforeEach, describe, expect, it, vi } from "vitest";
 import { App } from "./App";
-import { CommunityDetectionConfig, ControlMetrics, Dream, DreamRun, DreamStatus, DreamingConfig, GeneralConfig, OperationLog, OperationLogConfig, RecallFeedbackConfig, RecallFeedbackEvent, SecurityBan, SecuritySettings, SSOConfig, Team, TeamProfile } from "./api";
+import { CommunityDetectionConfig, ControlMetrics, Dream, DreamRun, DreamStatus, DreamingConfig, GeneralConfig, OperationLog, OperationLogConfig, SecurityBan, SecuritySettings, SSOConfig, Team, TeamProfile } from "./api";
 
 const profileA: Team = {
   id: "11111111-1111-4111-8111-111111111111",
@@ -144,18 +144,6 @@ const operationLogConfigSnapshot: OperationLogConfig = {
   },
 };
 
-const recallFeedbackConfigSnapshot: RecallFeedbackConfig = {
-  update_time: "2026-06-23T12:00:00Z",
-  items: [
-    { key: "RECALL_FEEDBACK_ENABLED", value: "true", effective_value: "true", updated_at: "2026-06-23T12:00:00Z" },
-    { key: "RECALL_FEEDBACK_RETENTION_DAYS", value: "30", effective_value: "30", updated_at: "2026-06-23T12:00:00Z" },
-  ],
-  effective: {
-    enabled: true,
-    retention_days: 30,
-  },
-};
-
 const communityDetectionConfigSnapshot: CommunityDetectionConfig = {
   update_time: "2026-06-15T03:30:00Z",
   items: [
@@ -213,59 +201,6 @@ const operationLogsSnapshot: OperationLog[] = [
       id_token_claim_count: 12,
       userinfo_claim_count: 11,
     },
-  },
-];
-
-const recallFeedbackEventsSnapshot: RecallFeedbackEvent[] = [
-  {
-    recall_id: "rec_1234567890",
-    created_at: "2026-06-23T10:00:00Z",
-    updated_at: "2026-06-23T10:01:00Z",
-    feedback_at: "2026-06-23T10:01:00Z",
-    team_id: profileA.id,
-    profile_id: keyA().id,
-    key_id: keyA().id,
-    auth_method: "api_key",
-    tool_name: "recall_memory",
-    query: "Why was recall bad?",
-    tool_args: {
-      input: { query: "Why was recall bad?", limit: 5 },
-      effective: { query: "Why was recall bad?", limit: 5, include_evidence: false, use_communities: false },
-    },
-    result_refs: [
-      {
-        type: "fragment",
-        id: "fragment-1",
-        rank: 1,
-        tier: "2",
-        final_score: 0.74,
-        status_at_recall: "active",
-      },
-    ],
-    result_count: 1,
-    snapshot_state: "captured",
-    used: true,
-    answer_supported: false,
-    quality: "low",
-    missing_context: true,
-    irrelevant: false,
-    resolved_results: [
-      {
-        type: "fragment",
-        id: "fragment-1",
-        rank: 1,
-        resolution_status: "found",
-        current_status: "retracted",
-        current: { content: "The old fragment has been retracted.", status: "retracted" },
-        ref: {
-          type: "fragment",
-          id: "fragment-1",
-          rank: 1,
-          tier: "2",
-          status_at_recall: "active",
-        },
-      },
-    ],
   },
 ];
 
@@ -603,31 +538,6 @@ describe("App", () => {
     });
   });
 
-  it("shows recall feedback query, params, result ids, and resolved state", async () => {
-    const fetchMock = mockPortalFetch({ teams: [profileA], keys: [keyA()] });
-    sessionStorage.setItem("denseMem.controlToken", "secret");
-
-    render(<App />);
-    await screen.findByRole("button", { name: /Default/ });
-    await userEvent.click(screen.getByRole("button", { name: /^recall feedback$/i }));
-
-    expect(await screen.findByText("Why was recall bad?")).toBeInTheDocument();
-    expect(screen.getByText("missing context")).toBeInTheDocument();
-
-    await userEvent.click(screen.getByRole("button", { name: /view recall feedback rec_1234567890/i }));
-    expect(await screen.findByText("fragment-1")).toBeInTheDocument();
-    expect(screen.getByText("The old fragment has been retracted.")).toBeInTheDocument();
-    expect(screen.getByLabelText(/Raw recall feedback rec_1234567890/i)).toHaveTextContent('"tool_args"');
-
-    await userEvent.selectOptions(screen.getByLabelText("Quality"), "low");
-    await waitFor(() => {
-      expect(fetchMock).toHaveBeenCalledWith(
-        expect.stringContaining("/recall-feedback-events?limit=100&offset=0&quality=low"),
-        expect.objectContaining({ method: "GET" }),
-      );
-    });
-  });
-
   it("shows dream rationale behind an info tooltip", async () => {
     mockPortalFetch({ teams: [profileA], keys: [keyA()] });
     sessionStorage.setItem("denseMem.controlToken", "secret");
@@ -664,32 +574,6 @@ describe("App", () => {
         expect.objectContaining({
           method: "PATCH",
           body: expect.stringContaining(`"SSO_PUBLIC_BASE_URL"`),
-        }),
-      );
-    });
-    expect(await screen.findByText("Saved")).toBeInTheDocument();
-  });
-
-  it("edits recall feedback retention config", async () => {
-    const fetchMock = mockPortalFetch({ teams: [profileA], keys: [keyA()] });
-    sessionStorage.setItem("denseMem.controlToken", "secret");
-
-    render(<App />);
-    await screen.findByRole("button", { name: /Default/ });
-    await userEvent.click(screen.getByRole("button", { name: /^Config$/i }));
-    await userEvent.click(within(await screen.findByRole("tablist", { name: /config sections/i })).getByRole("tab", { name: /^recall$/i }));
-
-    expect(await screen.findByRole("heading", { name: "Recall Feedback" })).toBeInTheDocument();
-    await userEvent.clear(screen.getByLabelText("Investigation retention days"));
-    await userEvent.type(screen.getByLabelText("Investigation retention days"), "45");
-    await userEvent.click(screen.getByRole("button", { name: /save config/i }));
-
-    await waitFor(() => {
-      expect(fetchMock).toHaveBeenCalledWith(
-        expect.stringContaining("/config/recall-feedback"),
-        expect.objectContaining({
-          method: "PATCH",
-          body: expect.stringContaining(`"RECALL_FEEDBACK_RETENTION_DAYS"`),
         }),
       );
     });
@@ -861,7 +745,6 @@ function mockPortalFetch({
   let currentDreamingConfig = structuredClone(dreamingConfigSnapshot);
   let currentCommunityDetectionConfig = structuredClone(communityDetectionConfigSnapshot);
   let currentOperationLogConfig = structuredClone(operationLogConfigSnapshot);
-  let currentRecallFeedbackConfig = structuredClone(recallFeedbackConfigSnapshot);
   const fetchMock = vi.fn(async (input: RequestInfo | URL, init?: RequestInit) => {
     const url = String(input);
     const method = init?.method ?? "GET";
@@ -949,34 +832,6 @@ function mockPortalFetch({
         }),
       };
       return jsonResponse({ data: currentOperationLogConfig });
-    }
-    if (url.endsWith("/config/recall-feedback") && method === "GET") {
-      return jsonResponse({ data: currentRecallFeedbackConfig });
-    }
-    if (url.endsWith("/config/recall-feedback") && method === "PATCH") {
-      const body = JSON.parse(String(init?.body));
-      currentRecallFeedbackConfig = {
-        ...currentRecallFeedbackConfig,
-        update_time: "2026-06-23T12:01:00Z",
-        items: currentRecallFeedbackConfig.items.map((item) => {
-          const update = body.items.find((candidate: { key: string }) => candidate.key === item.key);
-          return update ? { ...item, value: update.value, updated_at: "2026-06-23T12:01:00Z" } : item;
-        }),
-      };
-      return jsonResponse({ data: currentRecallFeedbackConfig });
-    }
-    if (parsedUrl.pathname.endsWith("/control/api/recall-feedback-events/rec_1234567890") && method === "GET") {
-      return jsonResponse({ data: recallFeedbackEventsSnapshot[0] });
-    }
-    if (parsedUrl.pathname.endsWith("/control/api/recall-feedback-events") && method === "GET") {
-      const limit = Number(parsedUrl.searchParams.get("limit") ?? "100");
-      const offset = Number(parsedUrl.searchParams.get("offset") ?? "0");
-      const quality = parsedUrl.searchParams.get("quality") ?? "";
-      const filtered = quality ? recallFeedbackEventsSnapshot.filter((event) => event.quality === quality) : recallFeedbackEventsSnapshot;
-      return jsonResponse({
-        data: filtered.slice(offset, offset + limit),
-        pagination: { limit, offset, total: filtered.length },
-      });
     }
     if (parsedUrl.pathname.endsWith("/control/api/logs") && method === "GET") {
       const limit = Number(parsedUrl.searchParams.get("limit") ?? "100");
